# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOtion API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tool {
  id: string;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  longDescription?: string | null;
  category: string;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  avgRating: number;
  totalReviews: number;
}

export interface Pagination {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}

export interface ToolListResponse {
  tools: Tool[];
  pagination: Pagination;
}

export interface ToolDetailResponse {
  tool: Tool;
}

export interface Review {
  id: string;
  toolSlug: string;
  authorUsername: string;
  /** @nullable */
  authorAvatar?: string | null;
  rating: number;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  upvotes: number;
}

export interface ReviewInput {
  toolSlug: string;
  authorUsername: string;
  authorAvatar?: string;
  rating: number;
  title: string;
  content: string;
}

export interface UpdateReviewInput {
  authorUsername: string;
  rating?: number;
  title?: string;
  content?: string;
}

export interface OwnerInput {
  authorUsername: string;
}

export interface VoterInput {
  voterUsername: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  pagination: Pagination;
}

export interface UpvoteResponse {
  upvotes: number;
}

export interface Category {
  slug: string;
  name: string;
  description: string;
  icon: string;
  toolCount: number;
}

export interface CategoryListResponse {
  categories: Category[];
}

export interface PlatformStats {
  totalTools: number;
  totalReviews: number;
  totalCategories: number;
}

export type ListToolsParams = {
category?: string;
page?: number;
limit?: number;
search?: string;
};

export type ListReviewsParams = {
toolSlug?: string;
page?: number;
limit?: number;
};