# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevOtion API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tools
    description: Tool operations
  - name: reviews
    description: Review operations
  - name: categories
    description: Category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List tools
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"

  /tools/{slug}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a tool by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolDetailResponse"
        "404":
          description: Not found

  /tools/{slug}/reviews:
    get:
      operationId: fetchToolReviews
      tags: [reviews]
      summary: List reviews for a tool
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews with optional filtering
      parameters:
        - name: toolSlug
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review (one per user per tool)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "409":
          description: Already reviewed this tool

  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found

    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update own review (owner-only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewInput"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "403":
          description: Forbidden — not the review author
        "404":
          description: Not found

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete own review (owner-only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerInput"
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden — not the review author
        "404":
          description: Not found

  /reviews/{id}/upvote:
    post:
      operationId: upvoteReview
      tags: [reviews]
      summary: Upvote a review (no self-votes, no duplicates)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterInput"
      responses:
        "200":
          description: Updated upvote count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpvoteResponse"
        "403":
          description: Forbidden — self-vote
        "409":
          description: Already upvoted

    delete:
      operationId: removeUpvote
      tags: [reviews]
      summary: Remove a previously cast upvote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoterInput"
      responses:
        "200":
          description: Updated upvote count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpvoteResponse"
        "404":
          description: Vote not found

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with tool counts
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"

  /stats:
    get:
      operationId: getStats
      tags: [tools]
      summary: Get platform stats
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Tool:
      type: object
      required: [id, name, slug, description, category, avgRating, totalReviews]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        category:
          type: string
        websiteUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        avgRating:
          type: number
        totalReviews:
          type: integer

    ToolListResponse:
      type: object
      required: [tools, pagination]
      properties:
        tools:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        pagination:
          $ref: "#/components/schemas/Pagination"

    ToolDetailResponse:
      type: object
      required: [tool]
      properties:
        tool:
          $ref: "#/components/schemas/Tool"

    Review:
      type: object
      required: [id, toolSlug, authorUsername, rating, title, content, createdAt, updatedAt, upvotes]
      properties:
        id:
          type: string
        toolSlug:
          type: string
        authorUsername:
          type: string
        authorAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        upvotes:
          type: integer

    ReviewInput:
      type: object
      required: [toolSlug, rating, title, content, authorUsername]
      properties:
        toolSlug:
          type: string
        authorUsername:
          type: string
        authorAvatar:
          type: string
        rating:
          type: integer
        title:
          type: string
        content:
          type: string

    UpdateReviewInput:
      type: object
      required: [authorUsername]
      properties:
        authorUsername:
          type: string
        rating:
          type: integer
        title:
          type: string
        content:
          type: string

    OwnerInput:
      type: object
      required: [authorUsername]
      properties:
        authorUsername:
          type: string

    VoterInput:
      type: object
      required: [voterUsername]
      properties:
        voterUsername:
          type: string

    ReviewListResponse:
      type: object
      required: [reviews, pagination]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        pagination:
          $ref: "#/components/schemas/Pagination"

    UpvoteResponse:
      type: object
      required: [upvotes]
      properties:
        upvotes:
          type: integer

    Category:
      type: object
      required: [slug, name, description, icon, toolCount]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        toolCount:
          type: integer

    CategoryListResponse:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"

    PlatformStats:
      type: object
      required: [totalTools, totalReviews, totalCategories]
      properties:
        totalTools:
          type: integer
        totalReviews:
          type: integer
        totalCategories:
          type: integer

    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer