# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOtion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  toolSlug: string;
  authorUsername: string;
  /** @nullable */
  authorAvatar?: string | null;
  rating: number;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  upvotes: number;
}