# lib/api-zod/src/generated/types/updateReviewInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * DevOtion API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateReviewInput {
  authorUsername: string;
  rating?: number;
  title?: string;
  content?: string;
}